Constants(marelac)	R Documentation

Useful physical and chemical constants

Description

Physical and chemical constants useful for marine and lacustrine sciences

Usage

Constants

Format

a list with, for each constant the value, the units, and a description.

Author(s)

Karline Soetaert <k.soetaert@nioo.knaw.nl>

References

to be filled...

Examples

data.frame(cbind(acronym=names(Constants),
              matrix(ncol=3,byrow=TRUE,data=unlist(Constants),
              dimnames=list(NULL,c("value","units","description")))))

[Package marelac version 1.1 Index]