O2sat(marelac)	R Documentation

Saturation concentration of oxygen in water

Description

Empirical formulae that can be used to compute saturation concentration of oxygen in water in mg L^{-1}

Usage


O2sat(S = 0, T = K - 273.15, K = T + 273.15, NN = 0, method = c("Weiss", "APHA", "Paul"))

Arguments

S 	 salinity in PSU (for method "Weiss" only),

T 	 Temperature in degrees centigrade,

K 	 absolute Temperature in degrees Kelvin,

NN 	 height above sea level (for method "Paul" only),

method 	 formula to be used, see references for correct application.

Details

Method APHA is the standard formula in Limnology, method Weiss possibly the standard formula in marine sciences. The method after Paul is a simple formula fitted on available tables. To avoid confusion between the arguments (S, T, K, NN) it is advisable to use named arguments in general (e.g. O2sat(T=4).

Value

Vector with oxygen saturation concentration in mg/L.

References

American Public Health Association, Inc. (1985): Standard Methods for the Examination of Water and Wastewater. 18th edition, 1992.

Benson B. B. and Krause, D. (1980): The concentration and isotopic fractionation of gases dissolved in freshwater in equilibrium with the atmosphere. I. Oxygen. Limnology and Oceanography 25, 662-671.

Brown, L.C. and Barnwell, T. O. Jr. (1987): The Enhanced Stream Water Quality Models QUAL2E and QUAL2E-UNCAS: Documentation and User Manual. U.S. Environmental Protection Agency, Athens, Georgia. EPA/600/3-87/007, p. 41. http://www.epa.gov)

DIN 38408-23, Ausgabe:1987-11: Deutsche Einheitsverfahren zur Wasser-, Abwasser- und Schlammuntersuchung; Gasf{"o}rmige Bestandteile (Gruppe G); Bestimmung des Sauerstoffs{"a}ttigungsindex (G 23). http://www.din.de

Paul, L. (1985). Das thermische Regime der Talsperre Saidenbach und einige Beziehungen zwischen abiotischen und biotischen Komponenten. Dissertation, TU Dresden, Fakult{"a}t Bau-, Wasser- und Forstwesen. 84 pp.

Weiss, R. (1970). The solubility of nitrogen, oxygen, and argon in water and seawater. Deep-Sea Res. 17: 721-35.

Wagner, R. (1979): Die Praxis der Bestimmung des biochemischen Sauerstoffbedarfs - Ergebnisse einer Umfrage (Berichtigung und Erg{"a}nzung zur Ver{"o}ffentlichung). Vom Wasser 53, S. 283-285.

See Also

satconc

Examples


  O2sat(T = 20)
  T <- seq(0, 30, 0.1)
  plot(T, O2sat(T = T), type = "l", ylim = c(0, 15), lwd=2)
  lines(T, O2sat(S = 0, T = T, method = "Weiss"), col = "yellow", lwd = 2, lty = "dashed")
  lines(T, O2sat(S = 35, T = T, method = "Weiss"), col = "red", lwd = 2)

[Package marelac version 1.3 Index]