marelac-package(marelac)	R Documentation

Datasets constants conversion factors and utilities for the marine and lacustrine sciences

Description

R-package marelac has been designed as a tool for use by scientists working in the marine and lacustrine sciences.

It contains:
data sets. These are mainly used in the lecture notes "Using R for scientific computing", written as an introduction to R for students that follow our course in ecological modelling.
chemical and physical constants, e.g. atomic weights, gas constants.
conversion factors, e.g. from salinity to chlorinity, from mol to gram, etc..
functions, e.g. to estimate concentrations of conservative substances as a function of salinity, ...

Details

Package: 	 marelac

Type: 	 Package

Version: 	 1.2

Date: 	 2008-11-031

License: 	 GNU Public License 2 or above

The package contains lecture notes for novice R-users. These have been written as an introduction to R , to prepare our students for the ecological modelling course (given by K. Soetaert and F. Meysman) in which R is used to implement and solve the models.

To open the lecture notes, copy-paste this sentence to R:

browseURL(paste(system.file(package="marelac"), "/lecture/UsingR.pdf", sep=""))

which will open a PDF generated with WORD (microsoft)

or type vignette("UsingR") which will open a LaTeX-generated file.

Author(s)

Karline Soetaert (Maintainer)

Filip Meysman

Examples


## Not run: 
## show examples (see respective help pages for details)
example(Bathymetry)
example(Constants)
example(SCOC)
example(mol.vol)

## open the directory with documents
browseURL(paste(system.file(package="marelac"), "/doc", sep=""))

## show document with the introduction to R - one file, a pdf generated with WORD:
browseURL(paste(system.file(package="marelac"), "/lecture/UsingR.pdf", sep=""))

## the introduction to R - a vignette
vignette("UsingR")

## and the answers
vignette("Answers")

edit(vignette("UsingR"))
edit(vignette("Answers"))

## End(Not run)

[Package marelac version 1.2 Index]