marelacTeaching-package(marelacTeaching)	R Documentation

Datasets and tutorials for use in the MArine Riverine Estuarine LAcustrine and Coastal sciences

Description

R-package marelacTeaching has been designed as a tool to help in teaching students from the MArine, Riverine, Estuarine, LAcustrine and Coastal sciences.

It contains:
a tutorial for novice R-users, focussing on the use of R in scientific computing.
data sets used in science teaching.

Details

Package: 	 marelac

Type: 	 Package

Version: 	 1.2

Date: 	 2008-11-031

License: 	 GNU Public License 2 or above

To open the lecture notes, copy-paste this sentence to R:

browseURL(paste(system.file(package="marelacTeaching"), "/lecture/UsingR.pdf", sep=""))

which will open a PDF generated with WORD (microsoft)

or type vignette("UsingR") which will open a LaTeX-generated file.

Author(s)

Karline Soetaert (Maintainer)

Thomas Petzoldt

Filip Meysman

Examples


## Not run: 
## show examples (see respective help pages for details)
example(Bathymetry)
example(SCOC)
example(Zoograzing)

## open the directory with documents
browseURL(paste(system.file(package="marelacTeaching"), "/doc", sep=""))

## show document with the introduction to R - one file, a pdf generated with WORD:
browseURL(paste(system.file(package="marelacTeaching"), "/lecture/UsingR.pdf", sep=""))

## the introduction to R - a vignette
vignette("UsingR")

## and the answers
vignette("Answers")

edit(vignette("UsingR"))
edit(vignette("Answers"))

## End(Not run)

[Package marelacTeaching version 1.0 Index]