mol.vol(marelac)	R Documentation

Molar volume of an ideal gas

Description

Calculates the molar volume of an ideal gas for given temperature and salinity

Usage

mol.vol(T=25, P=1)

Arguments

T 	 temperature, dg celsius

P 	 pressure, atmosphere

Value

molar volume, liter, of an ideal gas

Author(s)

Karline Soetaert <k.soetaert@nioo.knaw.nl>

See Also

mol2l and l2mol which converts from mol to liter and vice versa for any gas

Examples

mol.vol() # 24.46559
plot(0:30,mol.vol(T=0:30),xlab="Temperature, dgC",ylab="Molar volume")

[Package marelac version 1.1 Index]