redfield(marelac)	R Documentation

Redfield Ratio Calculator

Description

Estimate elemental composition of biomass (or media) according to the Redfield ratio.

Usage


redfield(element, x, method = c("mol", "mass"))

Arguments

element 	 The element that is given ("C", "H", "O", "N", "P"),

x 	 amount of substance of that element (in mol or mass units),

method 	 measurement unit ("mol" or "mass").

Details

The Redfield ratio is C106 H80 O45 N16 P1.

Value

Estimated ratio of the main elements.

References

Redfield A.C. (1934). On the proportions of organic derivations in sea water and their relation to the composition of plankton. In James Johnstone Memorial Volume. (ed. R.J. Daniel). University Press of Liverpool, 177-192.

See Also

mol.weight

Examples


# Redfield ratio
redfield("P", 1)
# returns the molar Redfield ratio, rescaled to nitrogen
redfield("N", 1)
# how many mass units are related to 2 mass units (e.g. mg) P
redfield("P", 2, "mass")

# mass percentage of elements
x <- redfield("P", 1, "mass")
x / sum(x)

# approximate fresh mass (in mg) formed by 1 microgram P
redfield("P", 1, "mass")["C"] * 2 * 10 / 1000


[Package marelac version 1.3 Index]