rhoH2O(marelac)	R Documentation

Density of Water

Description

Density of water in kg m^{-3}

Usage


rhoH2O(S=35, T=25, P=0, method=c("Millero", "Chen", "marine", "limnological"))

Arguments

S 	 salinity in PSU,

T 	 temperature in degree centigrade,

P 	 Hydrostatic pressure in bar (surface =0),

method 	 "Millero" for marine systems (default), "Chen" for the limnological range (i.e. fresh water systems). Abbreviation (e.g. "M" or "m") is possible.

Details

To avoid confusion between the arguments (S, T, P) it is advisable to use named arguments in general (e.g. rho(T=4, method="Chen"). The marine formula is imported from package seacarb.)

Value

Density of water in kg m^{-3}.

References

Chen, Ch.-T. and F.J. Millero (1986). Precise thermodynamic properties of natural waters covering only the limnological range. Limnol. Oceanogr. 31 No. 3, 657 - 662.

Fofonoff N.P. and R.C. Millard (1983). Algorithms for computation of fundamental properties of seawater. UNESCO technical papers in marine science, 44, 53 pp.

Millero, F. J. and Poisson, A. (1981). International one-atmosphere equation of state of seawater. Deep-Sea Research 28A, 625-629.

See Also

rho in package seacarb.

Examples


rhoH2O(T=4)
T <- 0:30
plot(T, rhoH2O(T=T))
S <- 0:40
plot(S, rhoH2O(S=S, T=4, method="Millero"))
lines(S, rhoH2O(S=S, T=4, method="Chen"))


[Package marelac version 1.3 Index]