salconc(marelac)	R Documentation

Concentrations of (conservative) species in seawater

Description

Estimates the concentration of Borate,Calcite, Sulphate and Fluoride in seawater

Usage

salconc(S=35)

Arguments

S 	 salinity

Details

The borate and calcite concentration as in Millero (1995)
Sulphate as in Morris and Riley, 1966
Fluoride as in Riley, 1965

Value

a data frame with the concentrations in micromol/kg

Author(s)

Karline Soetaert <k.soetaert@nioo.knaw.nl>

References

Millero, F.J., 1995. Thermodynamics of the carbon dioxide system in the oceans. Geochim. Cosmochim. Acta 59: 661 677.
Riley, J.P., 1965. The occurrence of anomalously high fluoride concentrations in the North Atlantic. Deep-Sea Res. 12: 219 220.
Morris, A.W., Riley, J.P., 1966. The bromide- chlorinity and sulphate- chlorinity ratio in seawater. Deep-Sea Res. 13: 699 706.

Examples

data.frame(salinity=1:35,salconc(1:35))

[Package marelac version 1.2 Index]