ssd2rad(marelac)	R Documentation

Estimate Global Radiation from Measured Sunshine Duration Time

Description

The function converts values of sunshine duration (in hours) to global radiation (in J cm^{-2} d^{-1})

Usage


ssd2rad(S, doy, a = 0.25, b = 0.5, rho = 50.29)

Arguments

S 	 Sunshine duration (hours)

doy 	 Julian day (for northern hemisphere only)

a, b, rho 	 site specific conversion parameters, must be fitted to measured data.

Value

Estimated value of global radiaton in J cm^{-2} d^{-1}.

Note

Don't forget to fit the function parameters to site specific values!

References

Dyck, S. u. G. Peschke (1995): Grundlagen der Hydrologie. 3. Auflage. Verlag f{"u}r Bauwesen, Berlin 1995, ISBN 3-345-00586-7

Examples


ssd2rad(8, 120)

[Package marelac version 1.3 Index]