vertmean(marelac)	R Documentation

Vertical Volumetric Weighted Mean of Matter Concentrations in Water Bodies

Description

Calculate vertical mean values which respect depths of different layers or lake morphometry.

Usage


vertmean(depth, vari, level, top, bot, vol)

Arguments

depth 	 sorted vector of sampled depths,

vari 	 measurements corresponding to depth (concentration, temperature, ...),

level 	 surface water level (above ground or m NN, depending on bathymetric function used,

top 	 top water level of requested layer,

bot 	 bottom water level of requested layer,

vol 	 hypsographic function to be used (e.g. vol.depth).

Value

Volumetric average value of the requested layer between depths top and bottom.

Examples


## define a bathymetric formula for a given lake or basin
weight.vol <- function(z, level) {
  ## Ground level 154m NN
  zz  <- level - 154 - z
  ## interpolate at very small values
  vol <- ifelse(zz < 0.5,
    approx(c(0, 0.5), c(0, 18020.74), zz)$y,
    ## Polynomial
    vol <- 44331 - 202176 * zz + 363281 * zz^2 - 150342 * zz^3 + 47275.8 * zz^4 -
          6791.125 * zz^5 + 503.9106 * zz^6 - 18.95684 * zz^7 + 0.2871075 * zz^8
  )
  vol
}

## dummy formula for depth-weighted averaging
## (water column, instead of bathymetric curve)
weight.column <- function(z, level) {z}

## Plot of bathymetric curve
x <- seq(0, 155, 0.01)
y <- weight.vol(0, x)
plot(x, y, xlim = c(154, max(x)))

## Test Data
level <- 167.5

depth <- c(0, 1, 6, 7, 8, 9, 10, 11)
pconc <- c(18.8, 17, 12.1, 11.9, 11.8, 11.7, 11.4, 11.4)

vertmean(depth, pconc, level, top=0, bot=4, weight.column)
vertmean(depth, pconc, level, top=4, bot=13.5, weight.column)

vertmean(depth, pconc, level, top=0, bot=4, weight.vol)
vertmean(depth, pconc, level, top=4, bot=13.5, weight.vol)

vertmean(depth, pconc, level, top=0, bot=3.7, weight.vol)
vertmean(depth, pconc, level, top=3.7, bot=13.5, weight.vol)

vertmean(depth, pconc, level, top=0, bot=3.7, weight.vol)
vertmean(depth, pconc, level, top=3.7, bot=13.5, weight.vol)


[Package marelac version 1.3 Index]